+++ src/Telegram/Adapter.ts
@@ -5,7 +5,6 @@ import request = require('request-promise-native');
 import {
     IAdapter,
     IResponse as IBotistResponse,
-    IError as IBotistError,
 } from '../Botist';
 import { IBaseMessage } from '../Message.t';
 
@@ -43,7 +42,7 @@ export class Telegram implements IAdapter {
         return [];
     }
 
-    public sendText(id: string, text: string): Promise<IBotistResponse | IBotistError> {
+    public sendText(id: string, text: string): Promise<IBotistResponse> {
         if (!text) {
             return Promise.resolve({ messageId: '' });
         }