+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased]
 
 ### Added
-- Exported types `IRequestResponse`, `IRequestError`, `IMessage`, `ITextMessage`, `IImageMessage`.
+- Exported types `IRequestResponse`, `IRequestSuccess`, `IRequestError`, `IMessage`, `ITextMessage`, `IImageMessage`.
 - Telegram/Adapter sendText() use Markdown parse mode.
 
 ## 0.2.1 (October 14, 2018)